Rename `query` to `contact_query`
